l3cfg: change l3cd for-each macros to take pointers instead of variable names
It's not a big difference and unclear which is preferable. It however
seems slighly clearer to require the user to provide a pointer, instead
of a variable for which the macro itself takes the reference.

It makes it clearer that the "iter" and "obj" arguments are modified
by the macro.