libnm: move nm_setting_ip4_config_get_dhcp_vendor_class_identifier() to "libnm_1_26_4"
nm_setting_ip4_config_get_dhcp_vendor_class_identifier() was
backported to nm-1-26 branch, and will be released as 1.26.4.
As such, on the stable branch the symbol will be placed in a
separate symbol version ("libnm_1_26_4").

To support the upgrade path from 1.26.4+ to 1.28+, we
want this symbol also present on master.

Signed-off-by: Antonio Cardace <acardace@redhat.com>