l3cfg/device: drop NML3Cfg references from NMDevice for now
Adding NML3Cfg support is a large effort that is done in parallel.
While already parts of the code is merged, it is not actually used
yet. Remove the parts from NMDevice that interact with NML3Cfg
before we actually start using it.

The point is that we might want to do a release before switching
over to the new way. For that release, we should not have the unused
code in NMDevice.

This patch will later be restored and extended.