libnm: relax asserting argument for nm_utils_hwaddr_len()
nm_utils_hwaddr_len() isn't very useful. It's documented to
only accept two possible input values (ARPHRD_ETHER and ARPHRD_INFINIBAND)
for which the respective return values are well known.

In particular, asserting that the input value is one of the two values
means it becomes harder to extend the function (and make it more useful).
Because, then the user would need to call:

     #if NM_VERSION > NM_VERSION_1_XX
     len = nm_utils_hwaddr_len (ARPHDR_FOO);
     #else
     len = 0
     #endif

and then it would introduce an unnecessary run time dependency on
NM_VERSION_1_XX.

Instead, just document to return 0 if the value is not supported.