platform/wifi: Use GET_STATION instead of GET_SCAN dumps
As an optimization, use the NL80211_CMD_GET_STATION dump data instead
of the NL80211_CMD_GET_SCAN dump + GET_STATION command (non-dump) to
implement the following methods:
wifi_nl80211_get_bssid
wifi_nl80211_get_rate
wifi_nl80211_get_qual

GET_STATION records vary in size from a few hundred bytes to a few kB.
GET_SCAN records are usually on the few hundred bytes side, but there
can be many of them.  In managed mode there will only be one
GET_STATION record.  In AdHoc mode there may be more.  These methods are
not used in AP or Mesh mode.

So without that patch we'd have a GET_SCAN dump that could be quite big
and then a GET_STATION with one record.  Now it should be a GET_STATION
dump with one record or a few records, in any case fewer synchronous
commands is better.  Additionally this should now not depend on the
currently-connected BSS being in the kernel's scan result cache.

The downside is that the signal strength is "optional" in the
GET_STATION records, depends on the driver's capabilities.  Most
mainline drivers do seem to include it (the mac80211 based ones and a
few full-mac ones) but I can't know if all of them do.